fix: error handling

+++ src/resources/boards/board.service.js
@@ -10,13 +10,16 @@ const addOne = board => boardRepo.addOne(board);
 const updateOne = (id, board) => boardRepo.updateOne(id, board);
 
 const deleteById = async boardId => {
-  await boardRepo.deleteById(boardId);
-  const tasks = await tasksRepo.getAll(boardId);
-  if (tasks) {
-    await Promise.all(
-      tasks.map(task => tasksRepo.deleteById(boardId, task.id))
-    );
+  const deletedCount = await boardRepo.deleteById(boardId);
+  if (deletedCount) {
+    const tasks = await tasksRepo.getAll(boardId);
+    if (tasks) {
+      await Promise.all(
+        tasks.map(task => tasksRepo.deleteById(boardId, task.id))
+      );
+    }
   }
+  return deletedCount;
 };
 
 module.exports = { getAll, getById, addOne, updateOne, deleteById };
